App is working again; added new cursor
Launch without error messages. Eraser is functional now and the curser is switched on circle while erasing. Font is bold for group titles. Gogol is removed (for this moment). Menu has submenus for Canvas button, but they don;t do anything yet (though the button functions are not empty). Need to be fixed